use the UINT64() macro instead of UL suffix, for 32-bit cleanliness
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@5367 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ dj/djsrpcc.cc
@@ -162,7 +162,7 @@ dj_rpc_call_gate(gate_sender *gs, time_t timeout,
 			  sys_clock_nsec() + retry_delivery_nsec);
 
 	while (jos_atomic_read(&rs.reply) == reply_copying)
-	    sys_sync_wait(&rs.reply.counter, reply_copying, ~0UL);
+	    sys_sync_wait(&rs.reply.counter, reply_copying, UINT64(~0));
 
 	if (jos_atomic_read(&rs.reply) == reply_done)
 	    break;
